requirements.txt
pandas
langchain-openai
langchain-experimental
tabulate
python-dotenv

